+++ Lib/Edge.h
@@ -9,7 +9,7 @@ namespace YinYangVis
 namespace Edge
 {
 
-    kvs::LineObject* CreateLineObject( const YinYangVis::YinYangVolumeObject* volume, const float dim_edge = 10.0f );
+    kvs::LineObject* CreateLineObject( const YinYangVis::YinYangVolumeObject* volume, const size_t dim_edge = 10 );
     kvs::LineObject* CreateLineObject( const YinYangVis::ZhongVolumeObject* volume );
 
 }
